Separate PyErrors from PConvert errors to allow conditional compilation of pyo3 code
