+++ Cargo.toml
@@ -10,6 +10,7 @@ crate-type = ["rlib", "cdylib"]
 
 [dependencies]
 image = "0.23.9"
+wasm-bindgen = "0.2.68"
 
 [dependencies.pyo3]
 version = "0.11.1"
