+++ src/lib.rs
@@ -2,8 +2,10 @@ mod benchmark;
 mod blending;
 mod constants;
 pub mod errors;
-mod pymodule;
 mod utils;
+
+#[cfg(all(target_arch = "wasm32", target_os = "unknown"))]
+mod pymodule;
 
 use benchmark::Benchmark;
 use blending::{
@@ -17,6 +19,18 @@ use std::env;
 use std::str;
 use std::str::FromStr;
 use utils::{read_png, write_png};
+
+use wasm_bindgen::prelude::*;
+
+#[wasm_bindgen]
+extern "C" {
+    pub fn alert(s: &str);
+}
+
+#[wasm_bindgen]
+pub fn greet(name: &str) {
+    alert(&format!("Hello, {}!", name));
+}
 
 pub fn pcompose(args: &mut env::Args) -> Result<(), PConvertError> {
     let dir = match args.next() {
